Initial commit
When importing console from loggy the stdout and stderr are taken over
and replaced with proxy objects that count the number of lines printed.
These objects also take care of making new lines hard when something
printed overflows the current line. The reason for this is that it's
really hard to count lines otherwise.

The line count is then used to keep track of things that have been
printed, as an example there is a very simple full-width progress bar
implemented that can be updated while things are printed underneeth it.